refactor: update login DTO structure
