+++ src/main/java/backend/knowhow/domain/auth/service/KakaoAuthService.java
@@ -1,44 +1,40 @@
 package backend.knowhow.domain.auth.service;
 
-import backend.knowhow.domain.auth.dto.KakaoUserInfo;
+import backend.knowhow.domain.auth.dto.response.KakaoUserInfo;
+import backend.knowhow.domain.auth.dto.response.KakaoUserResponse;
+import backend.knowhow.global.common.exception.BaseException;
+import backend.knowhow.global.common.response.ErrorType;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
-import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
 public class KakaoAuthService {
-    private final RestTemplate restTemplate = new RestTemplate();
 
+    private final RestTemplate restTemplate = new RestTemplate();
     public KakaoUserInfo getUserInfo(String accessToken) {
 
         HttpHeaders headers = new HttpHeaders();
         headers.set("Authorization", "Bearer " + accessToken);
 
         HttpEntity<Void> entity = new HttpEntity<>(headers);
-
-        ResponseEntity<Map> response = restTemplate.exchange(
-                "https://kapi.kakao.com/v2/user/me",
-                HttpMethod.GET,
-                entity,
-                Map.class
-        );
-
-        Map body = response.getBody();
-        Long id = ((Number) body.get("id")).longValue();
-
-        Map<String, Object> properties = (Map<String, Object>) body.get("properties");
-        Map<String, Object> account = (Map<String, Object>) body.get("kakao_account");
-
+        ResponseEntity<KakaoUserResponse> response =
+                restTemplate.exchange(
+                        "https://kapi.kakao.com/v2/user/me",
+                        HttpMethod.GET,
+                        entity,
+                        KakaoUserResponse.class
+                );
+        KakaoUserResponse body = response.getBody();
+        if (body == null || body.getId() == null) {
+            throw new BaseException(ErrorType.KAKAO_TOKEN_INVALID);
+        }
         return new KakaoUserInfo(
-                id,
-                (String) properties.get("nickname")
+                body.getId(),
+                body.getProperties().getNickname()
         );
     }
 }
